+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 2021-09-20
+
+Features:
+- More crawler options on input
+
+Bug fixes:
+- Re-patch SDK for 429 error
+- Don't use Chrome by default unless it's blocked
+- Don't close the browser midway the navigation
+
 ## 2021-09-10
 
 Features: